[GITHUB-3] Adds workflow details to readme
This will close #3. I have added some detail on how
to open PRs and contribute, how to control the versions
and some general notes on the structure of the roles.

I think this will be useful so we can just point to new
contributors to the readme rather than talking them
through it.